refactor(helper): move oceanbase-helper from dist out to root
