+++ internal/cmds/helper/env_check.go
@@ -10,14 +10,15 @@ MERCHANTABILITY OR FIT FOR A PARTICULAR PURPOSE.
 See the Mulan PSL v2 for more details.
 */
 
-package cmd
+package helper
 
 import (
 	"os"
 
 	"github.com/spf13/cobra"
 
-	"github.com/oceanbase/oceanbase-helper/pkg/oceanbase"
+	cmdconst "github.com/oceanbase/ob-operator/internal/const/cmd"
+	"github.com/oceanbase/ob-operator/pkg/helper"
 )
 
 func init() {
@@ -40,12 +41,12 @@ func newCheckStorageCmd() *cobra.Command {
 		Short: "Check whether the storage is ready",
 		Run: func(cmd *cobra.Command, args []string) {
 			if len(args) != 1 {
-				os.Exit(int(ExitCodeBadArgs))
+				os.Exit(int(cmdconst.ExitCodeBadArgs))
 			}
-			err := oceanbase.TryFallocate(args[0])
+			err := helper.TryFallocate(args[0])
 			if err != nil {
 				cmd.PrintErrf("Failed to check storage, %v \n", err)
-				os.Exit(int(ExitCodeErr))
+				os.Exit(int(cmdconst.ExitCodeNotSupport))
 			}
 		},
 	}
